+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## Unreleased
 
+- Added ability to open saved queries in a new GraphiQL tab instead of
+  overwriting the active tab.
 - Minor styling fixes
-
 
 ## 0.1.0 (2026-02-19)
 
 - Initial release.
-